% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleminusone.R
\name{doubleminusone}
\alias{doubleminusone}
\title{doubleminusone}
\usage{
doubleminusone(n)
}
\arguments{
\item{value}{}
}
\value{
value
}
\description{
Doubles a number and subtracts 1
}
\examples{
package:: doubleminusone(n)

}
